Rename: FileRootPathVO (domain.water_place -> domain)
